nexify-landing/training/hf_space/requirements.txt
# HF Space: Nexify Video Safety Assistant
streamlit>=1.40.0
torch>=2.0.0
transformers>=4.45.0
peft>=0.18.0
qwen-vl-utils>=0.0.14
pillow>=10.0.0
huggingface_hub>=0.26.0
